+++ frontend/src/css/layout.css
@@ -6,7 +6,7 @@
   display: flex;
   align-items: center;
   gap: 10px 30px;
-  margin: 10px 0;
+  margin: 20px 0;
   flex-wrap: wrap;
 }
